adding L3 neighbor delete test case
