+++ test/CMakeLists.txt
@@ -45,6 +45,7 @@ add_executable(
     gtest/aca_test_openflow.cpp
     gtest/aca_test_ovs_l2.cpp
     gtest/aca_test_ovs_l3.cpp
+    gtest/aca_test_ovs_util.cpp
 )
 
 # Link test executable against gtest & gtest_main
