+++ test/gtest/aca_test_ovs_l2.cpp
@@ -31,36 +31,42 @@ using aca_comm_manager::Aca_Comm_Manager;
 using aca_net_config::Aca_Net_Config;
 using aca_ovs_l2_programmer::ACA_OVS_L2_Programmer;
 
-string project_id = "99d9d709-8478-4b46-9f3f-000000000000";
-string vpc_id_1 = "1b08a5bc-b718-11ea-b3de-111111111111";
-string vpc_id_2 = "1b08a5bc-b718-11ea-b3de-222222222222";
-string subnet_id_1 = "27330ae4-b718-11ea-b3de-111111111111";
-string subnet_id_2 = "27330ae4-b718-11ea-b3de-222222222222";
-string port_id_1 = "01111111-b718-11ea-b3de-111111111111";
-string port_id_2 = "02222222-b718-11ea-b3de-111111111111";
-string port_id_3 = "03333333-b718-11ea-b3de-111111111111";
-string port_id_4 = "04444444-b718-11ea-b3de-111111111111";
-string port_name_1 = aca_get_port_name(port_id_1);
-string port_name_2 = aca_get_port_name(port_id_2);
-string port_name_3 = aca_get_port_name(port_id_3);
-string port_name_4 = aca_get_port_name(port_id_4);
-string vmac_address_1 = "fa:16:3e:d7:f2:6c";
-string vmac_address_2 = "fa:16:3e:d7:f2:6d";
-string vmac_address_3 = "fa:16:3e:d7:f2:6e";
-string vmac_address_4 = "fa:16:3e:d7:f2:6f";
-string vip_address_1 = "10.10.0.101";
-string vip_address_2 = "10.10.1.102";
-string vip_address_3 = "10.10.0.103";
-string vip_address_4 = "10.10.1.104";
-NetworkType vxlan_type = NetworkType::VXLAN;
-string subnet1_gw_ip = "10.10.0.1";
-string subnet2_gw_ip = "10.10.1.1";
-string subnet1_gw_mac = "fa:16:3e:d7:f2:11";
-string subnet2_gw_mac = "fa:16:3e:d7:f2:21";
-
+// extern the string and helper functions from aca_test_ovs_util.cpp
+extern string project_id;
+extern string vpc_id_1;
+extern string vpc_id_2;
+extern string subnet_id_1;
+extern string subnet_id_2;
+extern string port_id_1;
+extern string port_id_2;
+extern string port_id_3;
+extern string port_id_4;
+extern string port_name_1;
+extern string port_name_2;
+extern string port_name_3;
+extern string port_name_4;
+extern string vmac_address_1;
+extern string vmac_address_2;
+extern string vmac_address_3;
+extern string vmac_address_4;
+extern string vip_address_1;
+extern string vip_address_2;
+extern string vip_address_3;
+extern string vip_address_4;
 extern string remote_ip_1; // for docker network
 extern string remote_ip_2; // for docker network
+extern NetworkType vxlan_type;
+extern string subnet1_gw_ip;
+extern string subnet2_gw_ip;
+extern string subnet1_gw_mac;
+extern string subnet2_gw_mac;
 extern bool g_demo_mode;
+
+extern void aca_test_create_default_port_state(PortState *new_port_states);
+extern void aca_test_create_default_subnet_state(SubnetState *new_subnet_states);
+extern void aca_test_1_neighbor_CREATE_DELETE(NeighborType input_neighbor_type);
+extern void aca_test_1_port_CREATE_plus_neighbor_CREATE(NeighborType input_neighbor_type);
+extern void aca_test_10_neighbor_CREATE(NeighborType input_neighbor_type);
 
 // TODO: setup bridge when br-int is up and br-tun is gone
 
@@ -75,49 +81,6 @@ extern bool g_demo_mode;
 // TODO: neighbor invalid IP
 
 // TODO: neighbor invalid mac
-
-void aca_test_create_default_port_state(PortState *new_port_states)
-{
-  new_port_states->set_operation_type(OperationType::CREATE);
-
-  PortConfiguration *PortConfiguration_builder = new_port_states->mutable_configuration();
-  PortConfiguration_builder->set_revision_number(1);
-  PortConfiguration_builder->set_message_type(MessageType::FULL);
-  PortConfiguration_builder->set_id(port_id_1);
-
-  PortConfiguration_builder->set_project_id(project_id);
-  PortConfiguration_builder->set_vpc_id(vpc_id_1);
-  PortConfiguration_builder->set_name(port_name_1);
-  PortConfiguration_builder->set_mac_address(vmac_address_1);
-  PortConfiguration_builder->set_admin_state_up(true);
-
-  PortConfiguration_FixedIp *FixedIp_builder = PortConfiguration_builder->add_fixed_ips();
-  FixedIp_builder->set_subnet_id(subnet_id_1);
-  FixedIp_builder->set_ip_address(vip_address_1);
-
-  PortConfiguration_SecurityGroupId *SecurityGroup_builder =
-          PortConfiguration_builder->add_security_group_ids();
-  SecurityGroup_builder->set_id("1");
-}
-
-void aca_test_create_default_subnet_state(SubnetState *new_subnet_states)
-{
-  new_subnet_states->set_operation_type(OperationType::INFO);
-
-  SubnetConfiguration *SubnetConiguration_builder =
-          new_subnet_states->mutable_configuration();
-  SubnetConiguration_builder->set_revision_number(1);
-  SubnetConiguration_builder->set_project_id(project_id);
-  SubnetConiguration_builder->set_vpc_id(vpc_id_1);
-  SubnetConiguration_builder->set_id(subnet_id_1);
-  SubnetConiguration_builder->set_cidr("10.0.0.0/24");
-  SubnetConiguration_builder->set_tunnel_id(20);
-
-  auto *subnetConfig_GatewayBuilder(new SubnetConfiguration_Gateway);
-  subnetConfig_GatewayBuilder->set_ip_address(subnet1_gw_ip);
-  subnetConfig_GatewayBuilder->set_mac_address(subnet1_gw_mac);
-  SubnetConiguration_builder->set_allocated_gateway(subnetConfig_GatewayBuilder);
-}
 
 //
 // Test suite: ovs_l2_test_cases
@@ -261,7 +224,7 @@ TEST(ovs_l2_test_cases, 2_ports_CREATE_test_traffic_plus_neighbor_internal)
   overall_rc = EXIT_SUCCESS;
 }
 
-TEST(ovs_l2_test_cases, 1_port_CREATE_DELETE_demo_mode)
+TEST(ovs_l2_test_cases, 1_port_CREATE_DELETE)
 {
   ulong not_care_culminative_time = 0;
   string cmd_string;
@@ -319,190 +282,6 @@ TEST(ovs_l2_test_cases, 1_port_CREATE_DELETE_demo_mode)
 
   // free the allocated configurations since we are done with it now
   new_port_states->clear_configuration();
-  new_subnet_states->clear_configuration();
-
-  // delete br-int and br-tun bridges
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "del-br br-int", not_care_culminative_time, overall_rc);
-  EXPECT_EQ(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "del-br br-tun", not_care_culminative_time, overall_rc);
-  EXPECT_EQ(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-}
-
-TEST(ovs_l2_test_cases, 1_l2_neighbor_CREATE_DELETE)
-{
-  ulong not_care_culminative_time = 0;
-  string cmd_string;
-  int overall_rc;
-
-  // delete br-int and br-tun bridges
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "del-br br-int", not_care_culminative_time, overall_rc);
-
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "del-br br-tun", not_care_culminative_time, overall_rc);
-
-  // create and setup br-int and br-tun bridges, and their patch ports
-  overall_rc = ACA_OVS_L2_Programmer::get_instance().setup_ovs_bridges_if_need();
-  ASSERT_EQ(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-
-  GoalState GoalState_builder;
-  NeighborState *new_neighbor_states = GoalState_builder.add_neighbor_states();
-  SubnetState *new_subnet_states = GoalState_builder.add_subnet_states();
-
-  new_neighbor_states->set_operation_type(OperationType::CREATE);
-
-  // fill in neighbor state structs
-  NeighborConfiguration *NeighborConfiguration_builder =
-          new_neighbor_states->mutable_configuration();
-  NeighborConfiguration_builder->set_revision_number(1);
-
-  NeighborConfiguration_builder->set_project_id(project_id);
-  NeighborConfiguration_builder->set_vpc_id(vpc_id_1);
-  NeighborConfiguration_builder->set_id(port_id_3);
-  NeighborConfiguration_builder->set_mac_address(vmac_address_3);
-  NeighborConfiguration_builder->set_host_ip_address("172.17.111.222");
-
-  NeighborConfiguration_FixedIp *FixedIp_builder =
-          NeighborConfiguration_builder->add_fixed_ips();
-  FixedIp_builder->set_neighbor_type(NeighborType::L2);
-  FixedIp_builder->set_subnet_id(subnet_id_1);
-  FixedIp_builder->set_ip_address(vip_address_3);
-
-  // fill in subnet state structs
-  aca_test_create_default_subnet_state(new_subnet_states);
-
-  GoalStateOperationReply gsOperationalReply;
-
-  overall_rc = Aca_Comm_Manager::get_instance().update_goal_state(
-          GoalState_builder, gsOperationalReply);
-  ASSERT_EQ(overall_rc, EXIT_SUCCESS);
-
-  string outport_name = aca_get_outport_name(vxlan_type, "172.17.111.222");
-
-  // check if the outport has been created on br-tun
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          " list-ports br-tun | grep " + outport_name, not_care_culminative_time, overall_rc);
-  EXPECT_EQ(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-
-  // delete neighbor info
-  new_neighbor_states->set_operation_type(OperationType::DELETE);
-  overall_rc = Aca_Comm_Manager::get_instance().update_goal_state(
-          GoalState_builder, gsOperationalReply);
-  ASSERT_EQ(overall_rc, EXIT_SUCCESS);
-
-  // check if the outport has been deleted on br-tun
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          " list-ports br-tun | grep " + outport_name, not_care_culminative_time, overall_rc);
-  EXPECT_NE(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-
-  // clean up
-
-  // free the allocated configurations since we are done with it now
-  new_neighbor_states->clear_configuration();
-  new_subnet_states->clear_configuration();
-
-  // delete br-int and br-tun bridges
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "del-br br-int", not_care_culminative_time, overall_rc);
-  EXPECT_EQ(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "del-br br-tun", not_care_culminative_time, overall_rc);
-  EXPECT_EQ(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-}
-
-TEST(ovs_l2_test_cases, 1_port_CREATE_plus_neighbor_CREATE)
-{
-  ulong not_care_culminative_time = 0;
-  string cmd_string;
-  int overall_rc;
-
-  // delete br-int and br-tun bridges
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "del-br br-int", not_care_culminative_time, overall_rc);
-
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "del-br br-tun", not_care_culminative_time, overall_rc);
-
-  // create and setup br-int and br-tun bridges, and their patch ports
-  overall_rc = ACA_OVS_L2_Programmer::get_instance().setup_ovs_bridges_if_need();
-  ASSERT_EQ(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-
-  GoalState GoalState_builder;
-  PortState *new_port_states = GoalState_builder.add_port_states();
-  SubnetState *new_subnet_states = GoalState_builder.add_subnet_states();
-
-  // fill in port state structs
-  aca_test_create_default_port_state(new_port_states);
-
-  // fill in subnet state structs
-  aca_test_create_default_subnet_state(new_subnet_states);
-
-  // add a new neighbor state with CREATE
-  NeighborState *new_neighbor_states = GoalState_builder.add_neighbor_states();
-  new_neighbor_states->set_operation_type(OperationType::CREATE);
-
-  // fill in neighbor state structs
-  NeighborConfiguration *NeighborConfiguration_builder =
-          new_neighbor_states->mutable_configuration();
-  NeighborConfiguration_builder->set_revision_number(1);
-
-  NeighborConfiguration_builder->set_project_id(project_id);
-  NeighborConfiguration_builder->set_vpc_id(vpc_id_1);
-  NeighborConfiguration_builder->set_id(port_id_3);
-  NeighborConfiguration_builder->set_mac_address(vmac_address_3);
-  NeighborConfiguration_builder->set_host_ip_address("172.17.111.222");
-
-  NeighborConfiguration_FixedIp *FixedIp_builder =
-          NeighborConfiguration_builder->add_fixed_ips();
-  FixedIp_builder->set_neighbor_type(NeighborType::L2);
-  FixedIp_builder->set_subnet_id(subnet_id_1);
-  FixedIp_builder->set_ip_address(vip_address_3);
-
-  // set demo mode
-  bool previous_demo_mode = g_demo_mode;
-  g_demo_mode = true;
-
-  // configure the whole goal state in demo mode
-  GoalStateOperationReply gsOperationalReply;
-
-  overall_rc = Aca_Comm_Manager::get_instance().update_goal_state(
-          GoalState_builder, gsOperationalReply);
-  ASSERT_EQ(overall_rc, EXIT_SUCCESS);
-
-  // restore demo mode
-  g_demo_mode = previous_demo_mode;
-
-  // check to ensure the port 1 is created and setup correctly
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "get Interface " + port_name_1 + " ofport", not_care_culminative_time, overall_rc);
-  EXPECT_EQ(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-
-  string outport_name = aca_get_outport_name(vxlan_type, "172.17.111.222");
-
-  // check if the outport has been created on br-tun
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          " list-ports br-tun | grep " + outport_name, not_care_culminative_time, overall_rc);
-  EXPECT_EQ(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-
-  // clean up
-
-  // free the allocated configurations since we are done with it now
-  new_port_states->clear_configuration();
-  new_neighbor_states->clear_configuration();
   new_subnet_states->clear_configuration();
 
   // delete br-int and br-tun bridges
@@ -717,105 +496,19 @@ TEST(ovs_l2_test_cases, 10_ports_CREATE)
                gsOperationReply.message_total_operation_time() / 1000000);
 }
 
+TEST(ovs_l2_test_cases, 1_l2_neighbor_CREATE_DELETE)
+{
+  aca_test_1_neighbor_CREATE_DELETE(NeighborType::L2);
+}
+
+TEST(ovs_l2_test_cases, 1_port_CREATE_plus_l2_neighbor_CREATE)
+{
+  aca_test_1_port_CREATE_plus_neighbor_CREATE(NeighborType::L2);
+}
+
 TEST(ovs_l2_test_cases, 10_l2_neighbor_CREATE)
 {
-  string port_name_postfix = "-2222-3333-4444-555555555555";
-  string neighbor_id_postfix = "-baad-f00d-4444-555555555555";
-  string ip_address_prefix = "10.0.0.";
-  string remote_ip_address_prefix = "123.0.0.";
-  ulong not_care_culminative_time = 0;
-  int overall_rc;
-
-  // delete br-int and br-tun bridges
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "del-br br-int", not_care_culminative_time, overall_rc);
-
-  ACA_OVS_L2_Programmer::get_instance().execute_ovsdb_command(
-          "del-br br-tun", not_care_culminative_time, overall_rc);
-
-  // create and setup br-int and br-tun bridges, and their patch ports
-  overall_rc = ACA_OVS_L2_Programmer::get_instance().setup_ovs_bridges_if_need();
-  ASSERT_EQ(overall_rc, EXIT_SUCCESS);
-  overall_rc = EXIT_SUCCESS;
-
-  GoalState GoalState_builder;
-  NeighborState *new_neighbor_states;
-  SubnetState *new_subnet_states = GoalState_builder.add_subnet_states();
-  new_subnet_states->set_operation_type(OperationType::INFO);
-
-  SubnetConfiguration *SubnetConiguration_builder =
-          new_subnet_states->mutable_configuration();
-  SubnetConiguration_builder->set_revision_number(1);
-  SubnetConiguration_builder->set_project_id(project_id);
-  SubnetConiguration_builder->set_vpc_id("1b08a5bc-b718-11ea-b3de-111122223333");
-  SubnetConiguration_builder->set_id(subnet_id_1);
-  SubnetConiguration_builder->set_cidr("10.0.0.0/24");
-  SubnetConiguration_builder->set_tunnel_id(123);
-
-  auto *subnetConfig_GatewayBuilder(new SubnetConfiguration_Gateway);
-  subnetConfig_GatewayBuilder->set_ip_address(subnet1_gw_ip);
-  subnetConfig_GatewayBuilder->set_mac_address(subnet1_gw_mac);
-  SubnetConiguration_builder->set_allocated_gateway(subnetConfig_GatewayBuilder);
-
-  const int L2_NEIGHBOR_TO_CREATE = 10;
-
-  for (int i = 0; i < L2_NEIGHBOR_TO_CREATE; i++) {
-    string i_string = std::to_string(i);
-    string neighbor_id = i_string + neighbor_id_postfix;
-    string port_name = i_string + port_name_postfix;
-
-    new_neighbor_states = GoalState_builder.add_neighbor_states();
-    new_neighbor_states->set_operation_type(OperationType::CREATE);
-
-    NeighborConfiguration *NeighborConfiguration_builder =
-            new_neighbor_states->mutable_configuration();
-    NeighborConfiguration_builder->set_revision_number(1);
-
-    NeighborConfiguration_builder->set_project_id(project_id);
-    NeighborConfiguration_builder->set_vpc_id("1b08a5bc-b718-11ea-b3de-111122223333");
-    NeighborConfiguration_builder->set_id(neighbor_id);
-    NeighborConfiguration_builder->set_name(port_name);
-    NeighborConfiguration_builder->set_mac_address(vmac_address_1);
-    NeighborConfiguration_builder->set_host_ip_address(remote_ip_address_prefix + i_string);
-
-    NeighborConfiguration_FixedIp *FixedIp_builder =
-            NeighborConfiguration_builder->add_fixed_ips();
-    FixedIp_builder->set_neighbor_type(NeighborType::L2);
-    FixedIp_builder->set_subnet_id(subnet_id_1);
-    FixedIp_builder->set_ip_address(ip_address_prefix + i_string);
-  }
-
-  bool previous_demo_mode = g_demo_mode;
-  g_demo_mode = false;
-
-  GoalStateOperationReply gsOperationReply;
-  overall_rc = Aca_Comm_Manager::get_instance().update_goal_state(
-          GoalState_builder, gsOperationReply);
-  EXPECT_EQ(overall_rc, EXIT_SUCCESS);
-
-  g_demo_mode = previous_demo_mode;
-
-  // calculate the average latency
-  ulong total_neighbor_create_time = 0;
-
-  for (int i = 0; i < L2_NEIGHBOR_TO_CREATE; i++) {
-    ACA_LOG_DEBUG("Neighbor State(%d) took: %u nanoseconds or %u milliseconds\n",
-                  i, gsOperationReply.operation_statuses(i).state_elapse_time(),
-                  gsOperationReply.operation_statuses(i).state_elapse_time() / 1000000);
-
-    total_neighbor_create_time +=
-            gsOperationReply.operation_statuses(i).state_elapse_time();
-  }
-
-  ulong average_neighbor_create_time = total_neighbor_create_time / L2_NEIGHBOR_TO_CREATE;
-
-  ACA_LOG_INFO("Average L2 Neighbor Create of %d took: %lu nanoseconds or %lu milliseconds\n",
-               L2_NEIGHBOR_TO_CREATE, average_neighbor_create_time,
-               average_neighbor_create_time / 1000000);
-
-  ACA_LOG_INFO("[TEST METRICS] Elapsed time for message total operation took: %u nanoseconds or %u milliseconds\n",
-               gsOperationReply.message_total_operation_time(),
-               gsOperationReply.message_total_operation_time() / 1000000);
+  aca_test_10_neighbor_CREATE(NeighborType::L2);
 }
 
 TEST(ovs_l2_test_cases, DISABLED_2_ports_CREATE_test_traffic_PARENT)
